--- source/androidx.core/core/transforms/Metadata.ParameterNames.xml
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8" ?>
-<metadata>
-    <!--    
-    <attr 
-        path=""
-        name="managedName"
-        >
-        parameterName
-    </attr>
-    -->
-
-</metadata>
