--- source/com.android.support/support-compat/transforms/EnumFields.xml
@@ -1,7 +1,0 @@
-<enum-field-mappings>
-<!--  <mapping jni-class="android/support/v4/app/FragmentActivity$FragmentTag" clr-enum-type="Android.Support.V4.App.FragmentTagType">
-    <field  jni-name="Fragment_id"    clr-name="Id"   value="1" />
-    <field  jni-name="Fragment_name"  clr-name="Name" value="0" />
-    <field  jni-name="Fragment_tag"   clr-name="Tag"  value="2" />
-  </mapping>-->
-</enum-field-mappings>
